NodeSelectorContext supports potential and selected and convert/2
- NodeSelectorContext now aggregates matches(selected) nodes.
- NodeSelector.accept(Node, Consumer) replaced by accept(Node, NodeSelectorContext)
- Consumer replaced by NodeSelectorContext.potential.
- Removed NodeSelectorContext.asPredicate (impl as NodeSelectorNodeAttributeValuePredicate).
- Internal code cleanup.
- Created two selectors which accept Predicates, Predicate<Node>, Predicate<ExpressionEvaluationContext>
